requirements.txt
pandas
scikit-learn
joblib
streamlit
pyserial
